Reliably obtaining API name from handler
This fixes an issue when the request provided an URL path with two leading forward slashes (e.g. //api-docs/pets)  to the getApiDocumentation. This happened as the service was behind an Apache load balancing scheme that added an extra leading slash. Fixed by using AntHandlerMatcher to extract API name matched to the ** wildcard. I expect this to more reliably obtain the API name.