@ModelAttribute Request Params Now Expanded
Spring controller request params that are marked with @ModelAttribute
will now be flattened. In other words, instead of generating an
operation model that contains a complex object in the request, the
operation model will
appear to have been defined as accepting individual
simple type request parameters.

Using the @ModelAttribute on a controller method param
declares that Spring should bind request parameters to
the model object. As such, the Swagger model should
really show individual parameters, and not a complex
object type.