Fixed allowableValues not getting set for enum
Fixed bug where the allowableValues were not getting
set on enums when the enum property in a container
object had an ApiParam annotation. Apparently, if you
declare an ApiParam and don’t put any configuration in
for the allowableValues, the allowableValues will be a
blank String instead of null. The bug was that the code
was checking for a null allowableValues but not for a
blank String.