Provide support for overriding the request mappings
Both the swagger 1.2 and 2.0 controllers used constant endpoints
to access the swagger service description. This change adds a
property with a default value and provides a the PropertyPlaceholderConfigurer
post processor to provide the value at run-time and provides a
default value as well.

Also changed the endpoint for DefaultSwaggerController so that existing
clients will not break. Removed the v1 from the url.

fixes #169