Fixes the problem when multiple api listings are associated with the same api declaration
Swagger 1.2 does not support this model, hacked to only return the first api listing. This needs
to be fixed to merge the api listings by group, but thats for a different commit. This only
fixes the problem effectively for 2.0 of the spec, and spec 1.2 is no worse than it was before this
fix.

related to #727