Fixes a bug where bidirectional recursive types end up with missing properties
Before this change we ended up ignoring types that were seen. This should not be the case. We need to handle all the properties seen or unseen, so that properties are reflected correctly in the rendered schema. This change is so that we ignore ONLY the types that are ignorable.

(1196)