Reworked the RequestHandler
Converted the request handler to and interface. Also subclassed to accomodate work with web mvc types. This will allow for the spring data rest and grails controllers to participate.

Also tightend up the request mapping context and operation context so
that the the underlying request handlers do not leak beyond the
RequestHandler interface.