Distinguished models belonging to diffrent dockets
There was an oversight where the model context caching was not using the group name to distinguish the same model contexts from another docket. As a result models ignored in one docket and not the other would not behave as expected. Fixed this behavior and added tests to cover that use case.

fixes #1627