Fixed parameter type for params with multiple annoations
When a parameter is annotated with a RequestBody and a ModelAttribute we prioritize the model to appear as if it were a request body. When that happens we should not try and process the parameter as a model attribute. When this happens we log a warning and treat it as a body parameter regardless.

(2118)