Added logic to take the request path pattern into account
This is to address the case where there might be an endpoint with two request mapping one with the the parameter present in the url template and one without, in conjunction with the path variable marked as not required. As in the example below.

```java
@RequestMapping({"/foo", "/foo/{bar}"})
public void foo(@PathVariable(value = "bar", required = false) String bar)
...
```

Added additional tests to cover the different use cases for the path variables

(2220)