Remove the default sort.
This is just a modification with just two slash, but it means a lot. It fits a design principle that "do not modify something behind users". We known it intends to sort automatically for users. However it will confuse the user due to that the final order of dockets are different from the order specified by user. 
For example, the user defines three dockets with following order:
```
@Bean Docket1().....
@Bean Docket2().....
@Bean Docket3().....
```
But maybe the order shown in the home page become something below:
```
Docket3
Docket1
Docket2
```
This is confusing.
The best solution is removing this line of codes. Let the users decide the order by which dockets are shown.